Refactor Count metrics, and add QueueSeries
The original design was to imitate the same aggregation methods that the official
go DataDog client uses. However converting counts to rate is confusing, especially
as there is an actual count type in the API. Now there is a specific method for
count, and a specific method for rate.

Added the ability to queue DDMetric for the next flush.

Increased test coverage.