Consolidating main channel
Calls to Flush and Close are sent to separate channels then the
main metrics channel. This ensure that the signals are received
quickly, however this also creates a race condition. It's
possible to have a metric submitted first in a thread, followed
by a flush or a close, and have those different operations
processed out of order. To solve this, I'm consolidating
everything into a single channel.